Refactor code according to PR comments
Update uint32_t to size_t
Parse command line arguments to specify:
- max_message_size
- max_message_send
Rename ca_teec_check_result to ca_teec_print_result

+++ trusted_application_demo/ca_security_api/include/ca_security_api/aes_api.h
@@ -17,12 +17,12 @@ extern "C"
 TEEC_Result ca_aes_generate_encrypted_key(
   TEEC_Session * session,
   TEEC_SharedMemory * key,
-  uint32_t size_out);
+  size_t size_out);
 
 TEEC_Result ca_aes_decrypt_and_allocate_key(
   TEEC_Session * session,
   TEEC_SharedMemory * encrypted_key,
-  uint32_t size_encrypted_key);
+  size_t size_encrypted_key);
 
 # if __cplusplus
 }
